Extract AttachmentDraftStatusUpdater logic into a worker
