+++ app/services/service_listeners/attachment_draft_status_updater.rb
@@ -8,18 +8,7 @@ module ServiceListeners
 
     def update!
       return unless attachment_data.present?
-      draft = attachment_data.draft?
-      enqueue_job(attachment_data.file, draft)
-      if attachment_data.pdf?
-        enqueue_job(attachment_data.file.thumbnail, draft)
-      end
-    end
-
-  private
-
-    def enqueue_job(uploader, draft)
-      legacy_url_path = uploader.asset_manager_path
-      AssetManagerUpdateAssetWorker.perform_async(legacy_url_path, draft: draft)
+      AssetManagerAttachmentDraftStatusUpdateWorker.new.perform(attachment_data)
     end
   end
 end
